requirements.txt
# Agentic Journal - Python Dependencies

# Anthropic API
anthropic>=0.40.0

# YAML processing
pyyaml>=6.0

# PDF processing
pypdf>=4.0.0
pdfplumber>=0.10.0

# Markdown
markdown>=3.5

# HTTP
requests>=2.31.0

# CLI
click>=8.1.0

# Validation
jsonschema>=4.20.0

# Utils
python-dotenv>=1.0.0
